added viewSavedSearches functionality

+++ robot.py
@@ -6,14 +6,17 @@ from pyrosim.neuralNetwork import NEURAL_NETWORK
 import os
 import constants as c
 class ROBOT:
-    def __init__(self,solutionID):
+    def __init__(self,solutionID,nndfName = ""):
         self.robotId = p.loadURDF("body.urdf")
         self.solutionID = solutionID
         pyrosim.Prepare_To_Simulate(self.robotId)
         self.Prepare_To_Sense()
         self.Prepare_To_Act()
-        self.nn = NEURAL_NETWORK(f"brain{self.solutionID}.nndf")
-        os.system(f"rm brain{self.solutionID}.nndf")
+        if nndfName == "":
+            self.nn = NEURAL_NETWORK(f"brain{self.solutionID}.nndf")
+            os.system(f"rm brain{self.solutionID}.nndf")
+        else:
+            self.nn = NEURAL_NETWORK(nndfName)
 
     def Prepare_To_Sense(self):
         self.sensors = {}
